Fix patch_zipfile to account for equal paths
This fixes a bug where if the input and output paths are both
specified and point to the same file, the zipfile module would
raise an exception. The implemented solution checks if the paths
are the same and if so sets the output to None, causing a temp
file to be used and the original to ultimately be overwritten.